Moved OCI8Stmt into statement.go
